Extract AI run metadata helpers
